# public/assets/api.yaml
openapi: 3.0.0
paths:
...
  /sysRoleMenus:
    get:
      tags:
        - SysRoleMenu
      summary: 'Find list SysRoleMenu'
      description: 'Returns list of SysRoleMenu'
      operationId: getSysRoleMenu
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SysRoleMenu' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SysRoleMenu' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/SysRoleMenu' } }
                type: object
        '404':
          description: 'SysRoleMenu not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - SysRoleMenu
      summary: 'Add a new SysRoleMenu to the store'
      operationId: addSysRoleMenu
      requestBody:
        $ref: '#/components/requestBodies/SysRoleMenu'
      responses:
        '201':
          description: 'Created SysRoleMenu'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysRoleMenu'
            application/xml:
              schema:
                $ref: '#/components/schemas/SysRoleMenu'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/sysRoleMenus/{id}':
    get:
      tags:
        - SysRoleMenu
      summary: 'Find SysRoleMenu by ID'
      description: 'Returns a single SysRoleMenu'
      operationId: getSysRoleMenuById
      parameters:
        -
          name: id
          in: path
          description: 'ID of SysRoleMenu to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysRoleMenu'
            application/xml:
              schema:
                $ref: '#/components/schemas/SysRoleMenu'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'SysRoleMenu not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - SysRoleMenu
      summary: 'Update an existing SysRoleMenu'
      operationId: updateSysRoleMenu
      parameters:
        -
          name: id
          in: path
          description: 'SysRoleMenu id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/SysRoleMenu'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'SysRoleMenu not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - SysRoleMenu
      summary: 'Deletes a SysRoleMenu'
      operationId: deleteSysRoleMenu
      parameters:
        -
          name: id
          in: path
          description: 'SysRoleMenu id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
  /sysRolePermissions:
    get:
      tags:
        - SysRolePermission
      summary: 'Find list SysRolePermission'
      description: 'Returns list of SysRolePermission'
      operationId: getSysRolePermission
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SysRolePermission' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SysRolePermission' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/SysRolePermission' } }
                type: object
        '404':
          description: 'SysRolePermission not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - SysRolePermission
      summary: 'Add a new SysRolePermission to the store'
      operationId: addSysRolePermission
      requestBody:
        $ref: '#/components/requestBodies/SysRolePermission'
      responses:
        '201':
          description: 'Created SysRolePermission'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysRolePermission'
            application/xml:
              schema:
                $ref: '#/components/schemas/SysRolePermission'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/sysRolePermissions/{id}':
    get:
      tags:
        - SysRolePermission
      summary: 'Find SysRolePermission by ID'
      description: 'Returns a single SysRolePermission'
      operationId: getSysRolePermissionById
      parameters:
        -
          name: id
          in: path
          description: 'ID of SysRolePermission to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysRolePermission'
            application/xml:
              schema:
                $ref: '#/components/schemas/SysRolePermission'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'SysRolePermission not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - SysRolePermission
      summary: 'Update an existing SysRolePermission'
      operationId: updateSysRolePermission
      parameters:
        -
          name: id
          in: path
          description: 'SysRolePermission id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/SysRolePermission'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'SysRolePermission not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - SysRolePermission
      summary: 'Deletes a SysRolePermission'
      operationId: deleteSysRolePermission
      parameters:
        -
          name: id
          in: path
          description: 'SysRolePermission id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
...
components:
  schemas:
...
    SysRoleMenu:
      title: SysRoleMenu
      description: SysRoleMenu
      properties:
        role_id:
          nullable: false
          title: role_id
          description: role_id
          type: integer
          format: '-'
          maxLength: 20
        menu_id:
          nullable: false
          title: menu_id
          description: menu_id
          type: integer
          format: '-'
          maxLength: 20
      type: object
    SysRolePermission:
      title: SysRolePermission
      description: SysRolePermission
      properties:
        role_id:
          nullable: true
          title: role_id
          description: role_id
          type: integer
          format: '-'
          maxLength: 11
        permission_id:
          nullable: true
          title: permission_id
          description: permission_id
          type: integer
          format: '-'
          maxLength: 11
      type: object
...
  requestBodies:
...
    SysRoleMenu:
      description: 'SysRoleMenu object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/SysRoleMenu'
        application/xml:
          schema:
            $ref: '#/components/schemas/SysRoleMenu'
        application/json:
          schema:
            $ref: '#/components/schemas/SysRoleMenu'
    SysRolePermission:
      description: 'SysRolePermission object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/SysRolePermission'
        application/xml:
          schema:
            $ref: '#/components/schemas/SysRolePermission'
        application/json:
          schema:
            $ref: '#/components/schemas/SysRolePermission'
...